Add cog that syncs the guild on startup
